Added copyright license for all files.
Overall structure:
The NDMP server is split into different components which are subdirectories of ndmp - currently comm, utils, xdr. In each component the source code is in an src directory. comm is reponsible for actual client-server communication, utils provides thread-safe queues and hex dumps. XDR is incomplete as of now.

Issues (to be addressed in future commits):
1. Align code (Code seems to be aligned in local files, but is not in GitHub).
2. Support for spinlock locks.
3. Complete the XDR integration.